+++ database/migrations/2016_10_18_202643_create_incomes_table.php
@@ -16,7 +16,7 @@ class CreateIncomesTable extends Migration
 
             $table->increments('id');
 
-            $table->integer('amountReceived');
+            $table->decimal('amountReceived', 12, 2);
             $table->string('giver');
             $table->string('dateReceived');
             $table->string('receiptNumber');
